Make build and host envs the same for all packages where the build artifacts are re-used
The build and host environments are re-used for these three packages: the
artifacts are built in the top-level build.sh/bld.bat, and then copied into the
host environment in the individual outputs' install_llvm.sh/bat, and then
packaged.

In order to avoid errors, including having to rebuild the artifacts between
different outputs, these environment configurations need to be all equal.